# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Jobs",
        "description": "Laravel API documentation for the 'My Jobs' application.",
        "contact": {
            "email": "erick1souza1ago04@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/infos": {
            "post": {
                "tags": [
                    "Infos"
                ],
                "summary": "Create a new info",
                "description": "Create a new info with the provided data.",
                "operationId": "infos_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "description": "Code of the info.",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "info": {
                                        "description": "Information content.",
                                        "type": "string",
                                        "example": "14920388."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Info created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID of the created info.",
                                                    "type": "integer",
                                                    "example": "8"
                                                },
                                                "code": {
                                                    "description": "code of the created info.",
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "info": {
                                                    "description": "info of the created info.",
                                                    "type": "string",
                                                    "example": "14920388"
                                                },
                                                "user_id": {
                                                    "description": "user_id of the created info.",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "updated_at": {
                                                    "description": "updated_at of the created info.",
                                                    "type": "string",
                                                    "example": "2024-06-24T22:24:25.000000Z"
                                                },
                                                "created_at": {
                                                    "description": "created_at of the created info.",
                                                    "type": "string",
                                                    "example": "2024-06-24T22:24:25.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "There was a problem registering the info.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "There was a problem registering the info."
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/infos/{id}": {
            "put": {
                "tags": [
                    "Infos"
                ],
                "summary": "Update info",
                "description": "Update info with the provided data.",
                "operationId": "infos_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the info to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "description": "Code of the info.",
                                        "type": "string",
                                        "example": "3"
                                    },
                                    "info": {
                                        "description": "Information content.",
                                        "type": "string",
                                        "example": "Updated information content."
                                    },
                                    "user_id": {
                                        "description": "ID of the user updating the info.",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Info updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID of the updated info.",
                                                    "type": "integer",
                                                    "example": "8"
                                                },
                                                "code": {
                                                    "description": "Code of the updated info.",
                                                    "type": "string",
                                                    "example": "3"
                                                },
                                                "info": {
                                                    "description": "Information content of the updated info.",
                                                    "type": "string",
                                                    "example": "Updated information content."
                                                },
                                                "user_id": {
                                                    "description": "ID of the user who updated the info.",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "updated_at": {
                                                    "description": "Date and time of the last update.",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-24T22:30:45.000000Z"
                                                },
                                                "created_at": {
                                                    "description": "Date and time of creation.",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-24T22:24:25.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid info: unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid info: unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "There was a problem updating the info.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "There was a problem updating the info."
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}